requirements.txt
Django==3.1.4
djangorestframework==3.11.0
django-cors-headers==3.4.0
requests==2.11.1
gunicorn==20.0.4
urllib3==1.25.9